Reorganize constants in plans.
This will help make dependencies more focused, and avoid depending on
all of plans any time you need a constant.

Don't depend on plans for propTypes validation.

Update tests to match changes to plans.

Move another export to constants.
