+++ client/blocks/disconnect-jetpack/index.jsx
@@ -26,7 +26,7 @@ import {
 import { disconnect } from 'state/jetpack/connection/actions';
 import { setAllSitesSelected, navigate } from 'state/ui/actions';
 import { successNotice, errorNotice, infoNotice, removeNotice } from 'state/notices/actions';
-import { getPlanClass } from 'lib/plans/constants';
+import { getPlanClass } from 'lib/plans';
 import { getSiteSlug, getSiteTitle, getSitePlanSlug } from 'state/sites/selectors';
 import getRewindState from 'state/selectors/get-rewind-state';
