+++ client/blocks/plan-thank-you-card/index.jsx
@@ -17,10 +17,9 @@ import getRawSite from 'state/selectors/get-raw-site';
 import { getCurrentPlan } from 'state/sites/plans/selectors';
 import QuerySites from 'components/data/query-sites';
 import QuerySitePlans from 'components/data/query-site-plans';
-import { getPlan } from 'lib/plans';
+import { getPlan, getPlanClass } from 'lib/plans';
 import ThankYouCard from 'components/thank-you-card';
 import PlanIcon from 'components/plans/plan-icon';
-import { getPlanClass } from 'lib/plans/constants';
 
 /**
  * Style dependencies
