+++ client/blocks/upgrade-nudge-expanded/index.jsx
@@ -19,7 +19,8 @@ import PlanCompareCard from 'my-sites/plan-compare-card';
 import PlanCompareCardItem from 'my-sites/plan-compare-card/item';
 import TrackComponentView from 'lib/analytics/track-component-view';
 import { preventWidows } from 'lib/formatting';
-import { getFeatureTitle, getPlan } from 'lib/plans';
+import { getPlan } from 'lib/plans';
+import { getFeatureTitle } from 'lib/plans/features-list';
 import { getPlanBySlug } from 'state/plans/selectors';
 import { PLAN_PERSONAL } from 'lib/plans/constants';
 import { getSitePlan, getSiteSlug } from 'state/sites/selectors';
