+++ client/lib/plans/index.js
@@ -5,7 +5,7 @@
  */
 import moment from 'moment';
 import { format as urlFormat, parse as urlParse } from 'url';
-import { difference, get, includes, invoke, pick, values } from 'lodash';
+import { difference, get, includes, pick, values } from 'lodash';
 
 /**
  * Internal dependencies
@@ -13,8 +13,6 @@ import { difference, get, includes, invoke, pick, values } from 'lodash';
 import { isEnabled } from 'config';
 import { isFreeJetpackPlan, isJetpackPlan, isMonthly } from 'lib/products-values';
 import {
-	FEATURES_LIST,
-	PLANS_LIST,
 	PLAN_FREE,
 	PLAN_PERSONAL,
 	TERM_MONTHLY,
@@ -28,10 +26,8 @@ import {
 	TYPE_PREMIUM,
 	GROUP_WPCOM,
 	GROUP_JETPACK,
-	PLAN_MONTHLY_PERIOD,
-	PLAN_ANNUAL_PERIOD,
-	PLAN_BIENNIAL_PERIOD,
 } from './constants';
+import { PLANS_LIST } from './plans-list';
 
 /**
  * Module vars
@@ -51,24 +47,38 @@ export function getPlan( planKey ) {
 	return PLANS_LIST[ planKey ];
 }
 
-export function getValidFeatureKeys() {
-	return Object.keys( FEATURES_LIST );
-}
-
-export function isValidFeatureKey( feature ) {
-	return !! FEATURES_LIST[ feature ];
-}
-
-export function getFeatureByKey( feature ) {
-	return FEATURES_LIST[ feature ];
-}
-
-export function getFeatureTitle( feature ) {
-	return invoke( FEATURES_LIST, [ feature, 'getTitle' ] );
-}
-
 export function getPlanPath( plan ) {
-	return get( getPlan( plan ), 'getPathSlug', () => undefined )();
+	const retrievedPlan = getPlan( plan );
+	const slug = retrievedPlan.getPathSlug || ( () => undefined );
+	return slug();
+}
+
+export function getPlanClass( planKey ) {
+	if ( isFreePlan( planKey ) ) {
+		return 'is-free-plan';
+	}
+
+	if ( isBloggerPlan( planKey ) ) {
+		return 'is-blogger-plan';
+	}
+
+	if ( isPersonalPlan( planKey ) ) {
+		return 'is-personal-plan';
+	}
+
+	if ( isPremiumPlan( planKey ) ) {
+		return 'is-premium-plan';
+	}
+
+	if ( isBusinessPlan( planKey ) ) {
+		return 'is-business-plan';
+	}
+
+	if ( isEcommercePlan( planKey ) ) {
+		return 'is-ecommerce-plan';
+	}
+
+	return '';
 }
 
 /**
@@ -442,27 +452,4 @@ export function applyTestFiltersToPlansList( planName, abtest ) {
 	filteredPlanConstantObj.getPlanCompareFeatures = () => filteredPlanFeaturesConstantList;
 
 	return filteredPlanConstantObj;
-}
-
-/**
- * Return estimated duration of given PLAN_TERM in days
- *
- * @param {String} term TERM_ constant
- * @return {Number} Term duration
- */
-export function getTermDuration( term ) {
-	switch ( term ) {
-		case TERM_MONTHLY:
-			return PLAN_MONTHLY_PERIOD;
-
-		case TERM_ANNUALLY:
-			return PLAN_ANNUAL_PERIOD;
-
-		case TERM_BIENNIALLY:
-			return PLAN_BIENNIAL_PERIOD;
-	}
-
-	if ( process.env.NODE_ENV === 'development' ) {
-		console.error( `Unexpected argument ${ term }, expected one of TERM_ constants` ); // eslint-disable-line no-console
-	}
 }
