+++ client/my-sites/plan-features/header.jsx
@@ -19,8 +19,9 @@ import PlanPrice from 'my-sites/plan-price';
 import PlanIntervalDiscount from 'my-sites/plan-interval-discount';
 import PlanPill from 'components/plans/plan-pill';
 import PlanIcon from 'components/plans/plan-icon';
-import { TYPE_FREE, PLANS_LIST, getPlanClass } from 'lib/plans/constants';
-import { getYearlyPlanByMonthly, planMatches } from 'lib/plans';
+import { TYPE_FREE } from 'lib/plans/constants';
+import { PLANS_LIST } from 'lib/plans/plans-list';
+import { getYearlyPlanByMonthly, planMatches, getPlanClass } from 'lib/plans';
 import { getCurrentPlan } from 'state/sites/plans/selectors';
 import { getPlanBySlug } from 'state/plans/selectors';
 import { getSelectedSiteId } from 'state/ui/selectors';
