+++ client/state/plans/test/selectors.js
@@ -13,7 +13,9 @@ jest.mock( 'lib/plans/constants', () => ( {
 	TYPE_PREMIUM: 'TYPE_PREMIUM',
 	TYPE_BUSINESS: 'TYPE_BUSINESS',
 	TYPE_ECOMMERCE: 'TYPE_ECOMMERCE',
+} ) );
 
+jest.mock( 'lib/plans/plans-list', () => ( {
 	PLANS_LIST: {
 		jetpack_premium_monthly: {
 			term: 'TERM_MONTHLY',
