+++ client/state/products-list/selectors.js
@@ -13,7 +13,8 @@ import { pickBy, get, max } from 'lodash';
 import { abtest } from 'lib/abtest';
 import { getPlanDiscountedRawPrice } from 'state/sites/plans/selectors';
 import { getPlanRawPrice } from 'state/plans/selectors';
-import { getPlan, applyTestFiltersToPlansList, getTermDuration } from 'lib/plans';
+import { getPlan, applyTestFiltersToPlansList } from 'lib/plans';
+import { getTermDuration } from 'lib/plans/constants';
 
 export function isProductsListFetching( state ) {
 	return state.productsList.isFetching;
